exerciseapp_project/exercise_app/requirements.txt
asgiref==3.7.2
awsebcli==3.20.10
blessed==1.20.0
botocore==1.31.80
cement==2.8.2
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.3
Django==4.2.5
django-cors-headers==4.2.0
django-multiselectfield==0.1.12
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
pathspec==0.10.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
urllib3==1.26.18
wcwidth==0.1.9
whitenoise==6.4.0
